requirements.txt
importlib-resources==3.2.0
Pillow==8.0.1
pystray==0.17.1
python-xlib==0.28
six==1.15.0
